requirements.txt
asgiref==3.3.4
certifi==2020.12.5
Django==3.2.3
django-elasticsearch-dsl==7.2.0
djangorestframework==3.12.4
elasticsearch==7.12.1
elasticsearch-dsl==7.3.0
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.4
black==20.8b1
flake8==3.8.4
